Added README

+++ src/Counter.sol
@@ -5,10 +5,10 @@ contract Counter {
     uint256 public count;
 
     // Function to get the current count
-    function get() public view returns (uint) {
+    function get() public view returns (uint256) {
         return count;
     }
-    
+
     // Function to increment count by 1
     function inc() external {
         count += 1;
